backend/requirements.txt
flask
flask_cors
Flask-DotEnv
SQLAlchemy
PyJWT
